+++ node/src/main/kotlin/net/corda/node/services/network/PersistentNetworkMapCache.kt
@@ -3,6 +3,7 @@ package net.corda.node.services.network
 import net.corda.core.concurrent.CordaFuture
 import net.corda.core.crypto.SecureHash
 import net.corda.core.crypto.toStringShort
+import net.corda.core.identity.AbstractParty
 import net.corda.core.identity.CordaX500Name
 import net.corda.core.identity.Party
 import net.corda.core.identity.PartyAndCertificate
@@ -11,15 +12,17 @@ import net.corda.core.internal.concurrent.openFuture
 import net.corda.core.messaging.DataFeed
 import net.corda.core.node.NodeInfo
 import net.corda.core.node.NotaryInfo
+import net.corda.core.node.services.IdentityService
 import net.corda.core.node.services.NetworkMapCache.MapChange
 import net.corda.core.node.services.PartyInfo
 import net.corda.core.serialization.SingletonSerializeAsToken
 import net.corda.core.serialization.serialize
 import net.corda.core.utilities.NetworkHostAndPort
+import net.corda.core.utilities.Try
 import net.corda.core.utilities.contextLogger
 import net.corda.core.utilities.debug
 import net.corda.node.internal.schemas.NodeInfoSchemaV1
-import net.corda.node.services.api.NetworkMapCacheBaseInternal
+import net.corda.node.services.api.NetworkMapCacheInternal
 import net.corda.node.utilities.NonInvalidatingCache
 import net.corda.nodeapi.internal.persistence.CordaPersistence
 import net.corda.nodeapi.internal.persistence.bufferUntilDatabaseCommit
@@ -34,7 +37,8 @@ import javax.annotation.concurrent.ThreadSafe
 /** Database-based network map cache. */
 @ThreadSafe
 open class PersistentNetworkMapCache(private val database: CordaPersistence,
-                                     private val myLegalName: CordaX500Name) : SingletonSerializeAsToken(), NetworkMapCacheBaseInternal {
+                                     private val identityService: IdentityService,
+                                     private val myLegalName: CordaX500Name) : NetworkMapCacheInternal, SingletonSerializeAsToken() {
     companion object {
         private val logger = contextLogger()
     }
@@ -75,6 +79,15 @@ open class PersistentNetworkMapCache(private val database: CordaPersistence,
         }
         if (otherNodeInfoCount > 0) {
             _nodeReady.set(null)
+        }
+    }
+
+    override fun getNodeByLegalIdentity(party: AbstractParty): NodeInfo? {
+        return database.transaction {
+            val wellKnownParty = identityService.wellKnownPartyFromAnonymous(party)
+            wellKnownParty?.let {
+                getNodesByLegalIdentityKey(it.owningKey).firstOrNull()
+            }
         }
     }
 
@@ -160,6 +173,7 @@ open class PersistentNetworkMapCache(private val database: CordaPersistence,
             val previousNode = getNodesByLegalIdentityKey(node.legalIdentities.first().owningKey).firstOrNull()
             if (previousNode == null) {
                 logger.info("No previous node found")
+                if (!verifyAndRegisterIdentities(node)) return
                 database.transaction {
                     updateInfoDB(node, session)
                     changePublisher.onNext(MapChange.Added(node))
@@ -169,6 +183,8 @@ open class PersistentNetworkMapCache(private val database: CordaPersistence,
                 return
             } else if (previousNode != node) {
                 logger.info("Previous node was found as: $previousNode")
+                // TODO We should be adding any new identities as well
+                if (!verifyIdentities(node)) return
                 database.transaction {
                     updateInfoDB(node, session)
                     changePublisher.onNext(MapChange.Modified(node, previousNode))
@@ -181,6 +197,27 @@ open class PersistentNetworkMapCache(private val database: CordaPersistence,
             _nodeReady.set(null)
         }
         logger.debug { "Done adding node with info: $node" }
+    }
+
+    private fun verifyIdentities(node: NodeInfo): Boolean {
+        val failures = node.legalIdentitiesAndCerts.mapNotNull { Try.on { it.verify(identityService.trustAnchor) } as? Try.Failure }
+        if (failures.isNotEmpty()) {
+            logger.warn("$node has ${failures.size} invalid identities:")
+            failures.forEach { logger.warn("", it) }
+        }
+        return failures.isEmpty()
+    }
+
+    private fun verifyAndRegisterIdentities(node: NodeInfo): Boolean {
+        // First verify all the node's identities are valid before registering any of them
+        return if (verifyIdentities(node)) {
+            for (identity in node.legalIdentitiesAndCerts) {
+                identityService.verifyAndRegisterIdentity(identity)
+            }
+            true
+        } else {
+            false
+        }
     }
 
     override fun removeNode(node: NodeInfo) {
